move crawlers from the merge to a separate directory
move crawlers from the merge to a separate directory
