+++ crawlers/isekaiscanmanga.py
