+++ crawlers/kakalotmanga.py
